Renamed IResourceCollection to IResourceMap Made IResourceMap generic so that it can replace ILinkCollection and IChainCollection Created IAggregate interface Created IName interface and changed IResource to include an IName instead of a string

+++ trunk/alxlib/src/Libraries/Babel/IChain.cs
@@ -5,7 +5,8 @@ using System.Text;
 
 namespace Babel
 {
-	public interface IChain : IResource
+	public interface IChain :
+		IResource
 	{
 		IEnumerable<T> GetValues<T>()
 			where T : struct;
@@ -25,6 +26,6 @@ namespace Babel
 		void SetResources<T>(IEnumerable<T> resources)
 			where T : IResource;
 		
-		long GetCount();
+		uint GetCount();
 	}
 }
